==> notebooks/1c. Scrape BGT data.ipynb <==
# Scrape and parse reference data

This notebook helps to scrape the necessary reference data (mainly BGT-like data in the form of points and polygons).

# Add project src to path.
import set_path

import pathlib
import pandas as pd
import geopandas as gpd

from upcp.utils import las_utils

import upc_sw.poly_utils as poly_utils
import upc_sw.scraping_utils as scraping

### SETTINGS ###

# Specify individual tiles to scrape data for
tiles = ['2386_9702']

# Or, use all tiles in a given folder
pc_folder = '../datasets/pointcloud/'
# tiles = las_utils.get_tilecodes_from_folder(pc_folder)

# Output file for the BGT fuser.
bgt_folder = '../datasets/bgt/'

# Create folder if it does not exist
pathlib.Path(bgt_folder).mkdir(parents=True, exist_ok=True)

## 0. Prepare shape (optional)

To scrape BGT for specific shapes based on neighbourhoods, use this snippet.

# For neighbourhoods only: specify which neighbourhoods to keep
neighbourhood_data = '../datasets/neighbourhoods_amsterdam.csv'
neighbourhood_codes = ['M27', 'M28']

# Load neighbourhood data and convert to RD
neighbourhood_gdf = pd.read_csv(neighbourhood_data, index_col=0)
neighbourhood_gdf['geometry'] = gpd.GeoSeries.from_wkt(neighbourhood_gdf['geometry'])
neighbourhood_gdf = gpd.GeoDataFrame(neighbourhood_gdf, crs='wgs84')
neighbourhood_gdf.to_crs(epsg='28992', inplace=True)

# Filter by target areas
neighbourhood_gdf['target'] = neighbourhood_gdf['neighbourhood_combination_code'].isin(neighbourhood_codes)

# Merge areas
target_area = neighbourhood_gdf[neighbourhood_gdf['target']].unary_union

(x_min, y_min, x_max, y_max) = target_area.bounds
bbox = ((x_min, y_max), (x_max, y_min))

## 1. Sidewalk polygons

We first scrape sidewalk ("voetpad") polygons from the Amsterdam BGT API. The documentation of this data source can be found at:
- https://www.amsterdam.nl/stelselpedia/bgt-index/producten-bgt/prodspec-bgt-dgn-imgeo

# Output file
bgt_data_file = f'{bgt_folder}bgt_voetpad.gpkg'

# BGT layers and output headers
bgt_layers = ['BGT_WGL_voetpad']

# Process single tile or list of tiles
gdf = scraping.process_tiles(tiles, bgt_layers)

# Or, scrape an area based on all files in a folder
gdf = scraping.process_folder(pc_folder, bgt_layers)

# Or, scrape a shape
gdf = scraping.get_bgt_data_for_bbox(bbox, bgt_layers)
gdf = gdf[gdf.intersects(target_area)]

# Save as GeoPackage file
gdf.to_file(bgt_data_file, driver='GPKG')

## 2. "Terras" polygons

We use these to mark "horeca" terraces as static obstacles.

See https://data.amsterdam.nl/datasets/GsY50tEkoJKCGw/

# Output file
terras_data_file = f'{bgt_folder}terras_data.gpkg'

# Process single tile or list of tiles
gdf = scraping.process_tiles(tiles, None, scraper=scraping.get_terras_data_for_bbox)

# Or, scrape an area based on all files in a folder
gdf = scraping.process_folder(pc_folder, None, scraper=scraping.get_terras_data_for_bbox)

# Or, scrape a shape
gdf = scraping.get_terras_data_for_bbox(bbox)
gdf = gdf[gdf.intersects(target_area)]

# Save as GeoPackage file
gdf.to_file(terras_data_file, driver='GPKG')

## 3. BGT obstacle points

For now only trees.

# Output file
obstacle_data_file = f'{bgt_folder}obstacle_data.gpkg'

# BGT layers and output headers
bgt_layers = ['BGTPLUS_VGT_boom']

# Process single tile or list of tiles
gdf = scraping.process_tiles(tiles, bgt_layers)

# Or, scrape an area based on all files in a folder
gdf = scraping.process_folder(pc_folder, bgt_layers)

# Or, scrape a shape
gdf = scraping.get_bgt_data_for_bbox(bbox, bgt_layers)
gdf = gdf[gdf.intersects(target_area)]

# Save as GeoPackage file
gdf.to_file(obstacle_data_file, driver='GPKG')
